Refactor and redesign the front end
This commit changes the design and starts to make use of the new `description` field in the api response, to show a second panel containing the text of the job being advertised. Also switched out the CSS for bootstrap to get some grid layout and easier formatting to get rolling on this.